All commands fixed. mojang command system completed
