+++ src/main/java/thunder/hack/cmd/Command.java
@@ -1,44 +1,65 @@
 package thunder.hack.cmd;
 
-import thunder.hack.Thunderhack;
+import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.ArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.mojang.brigadier.builder.RequiredArgumentBuilder;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.command.CommandRegistryAccess;
+import net.minecraft.command.CommandSource;
+import net.minecraft.registry.BuiltinRegistries;
+import net.minecraft.server.command.CommandManager;
 import net.minecraft.text.Text;
+import thunder.hack.Thunderhack;
+
+import java.util.Arrays;
+import java.util.List;
 
 import static thunder.hack.modules.Module.fullNullCheck;
 
 public abstract class Command {
-    protected String name;
+    protected static final CommandRegistryAccess REGISTRY_ACCESS = CommandManager.createRegistryAccess(BuiltinRegistries.createWrapperLookup());
+    protected static final MinecraftClient MC = MinecraftClient.getInstance();
 
-    public static MinecraftClient mc;
+    protected final List<String> names;
 
-    public Command(String name) {
-        this.name = name;
-        mc = MinecraftClient.getInstance();
+    public Command(String... names) {
+        this.names = Arrays.asList(names);
     }
 
+    public abstract void executeBuild(LiteralArgumentBuilder<CommandSource> builder);
+
     public static void sendMessage(String message) {
-        Command.sendSilentMessage(Thunderhack.commandManager.getClientMessage() + " "  + message);
+        sendSilentMessage(Thunderhack.commandManager.getClientMessage() + " "  + message);
     }
 
     public static void sendMessageWithoutTH(String message) {
-        Command.sendSilentMessage(message);
+        sendSilentMessage(message);
     }
 
     public static void sendSilentMessage(String message) {
-        if (fullNullCheck()) {
-            return;
+        if (fullNullCheck()) return;
+
+        MC.player.sendMessage(Text.of(message));
+    }
+
+    protected static <T> RequiredArgumentBuilder<CommandSource, T> arg(final String name, final ArgumentType<T> type) {
+        return RequiredArgumentBuilder.argument(name, type);
+    }
+
+    protected static LiteralArgumentBuilder<CommandSource> literal(final String name) {
+        return LiteralArgumentBuilder.literal(name);
+    }
+
+    public void register(CommandDispatcher<CommandSource> dispatcher) {
+        for (String name : names) {
+            LiteralArgumentBuilder<CommandSource> builder = LiteralArgumentBuilder.literal(name);
+            executeBuild(builder);
+            dispatcher.register(builder);
         }
-        mc.player.sendMessage(Text.of(message));
     }
-
-    public static String getCommandPrefix() {
-        return Thunderhack.commandManager.getPrefix();
-    }
-
-    public abstract void execute(String[] var1);
 
     public String getName() {
-        return this.name;
+        return names.get(0);
     }
 }
-
