+++ src/main/java/thunder/hack/cmd/args/ModuleArgumentType.java
@@ -1,20 +1,19 @@
-package thunder.hack.command.args;
+package thunder.hack.cmd.args;
 
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.arguments.ArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
-import com.mojang.brigadier.suggestion.Suggestion;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
+import net.minecraft.command.CommandSource;
 import net.minecraft.text.Text;
 import thunder.hack.Thunderhack;
 import thunder.hack.modules.Module;
 import thunder.hack.modules.client.MainSettings;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public class ModuleArgumentType implements ArgumentType<Module> {
@@ -32,6 +31,11 @@ public class ModuleArgumentType implements ArgumentType<Module> {
         ).create(reader.readString());
 
         return module;
+    }
+
+    @Override
+    public <S> CompletableFuture<Suggestions> listSuggestions(CommandContext<S> context, SuggestionsBuilder builder) {
+        return CommandSource.suggestMatching(Thunderhack.moduleManager.modules.stream().map(Module::getName), builder);
     }
 
     @Override
