+++ src/main/java/thunder/hack/cmd/impl/CfgCommand.java
@@ -1,65 +1,52 @@
 package thunder.hack.cmd.impl;
 
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.minecraft.command.CommandSource;
 import thunder.hack.Thunderhack;
 import thunder.hack.cmd.Command;
-import thunder.hack.core.ConfigManager;
-import thunder.hack.modules.client.MainSettings;
+import thunder.hack.cmd.args.CfgArgumentType;
 
-import java.awt.*;
 import java.io.File;
 import java.util.Objects;
 
+import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
+
 public class CfgCommand extends Command {
     public CfgCommand() {
-        super("cfg");
+        super("cfg", "config");
     }
 
-    public void execute(String[] commands) {
-        if (commands.length == 1) {
+    @Override
+    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
+        builder.then(literal("list").executes(context -> {
+            StringBuilder configs = new StringBuilder("Configs: ");
+            for (String str : Objects.requireNonNull(Thunderhack.configManager.getConfigList())) {
+                configs.append("\n- ").append(str);
+            }
+            sendMessage(configs.toString());
 
-            if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                sendMessage("Конфиги сохраняются в  ThunderHackRecode/configs/");
-            } else {
-                sendMessage("Configurations are saved in ThunderHackRecode/configs/");
+            return SINGLE_SUCCESS;
+        }));
+
+        builder.then(literal("dir").executes(context -> {
+            try {
+                net.minecraft.util.Util.getOperatingSystem().open(new File("ThunderHackRecode/configs/").toURI());
+            } catch (Exception e) {
+                e.printStackTrace();
             }
-            return;
-        }
-        if (commands.length == 2)
-            if ("list".equals(commands[0])) {
-                StringBuilder configs = new StringBuilder("Configs: ");
-                for(String str : Objects.requireNonNull(Thunderhack.configManager.getConfigList())){
-                    configs.append("\n- ").append(str);
-                }
-                sendMessage(configs.toString());
-            } else if( "dir".equals(commands[0]) ){
-                try {
-                    net.minecraft.util.Util.getOperatingSystem().open(new File("ThunderHackRecode/configs/").toURI());
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            } else {
-                if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                    sendMessage("Нет такой команды!... Может list ?");
-                } else {
-                    sendMessage("Wrong command!... Maybe list?");
-                }
-            }
-        if (commands.length >= 3) {
-            switch (commands[0]) {
-                case "save":
-                case "create":
-                    Thunderhack.configManager.save(commands[1]);
-                    return;
-                case "set":
-                case "load":
-                    Thunderhack.configManager.load(commands[1]);
-                    return;
-            }
-            if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                sendMessage("Нет такой команды! Пример использования: <save/load>");
-            } else {
-                sendMessage("Wrong command! try: <save/load>");
-            }
-        }
+
+            return SINGLE_SUCCESS;
+        }));
+
+        builder.then(literal("save").then(arg("name", StringArgumentType.word()).executes(context -> {
+            Thunderhack.configManager.save(context.getArgument("name", String.class));
+            return SINGLE_SUCCESS;
+        })));
+
+        builder.then(literal("load").then(arg("name", CfgArgumentType.create()).executes(context -> {
+            Thunderhack.configManager.load(context.getArgument("name", String.class));
+            return SINGLE_SUCCESS;
+        })));
     }
 }
