+++ src/main/java/thunder/hack/cmd/impl/DrawCommand.java
@@ -1,10 +1,14 @@
 package thunder.hack.cmd.impl;
 
-import thunder.hack.Thunderhack;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.minecraft.command.CommandSource;
+import net.minecraft.util.Formatting;
 import thunder.hack.cmd.Command;
+import thunder.hack.cmd.args.ModuleArgumentType;
 import thunder.hack.gui.clickui.ClickUI;
 import thunder.hack.modules.Module;
-import net.minecraft.util.Formatting;
+
+import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 
 public class DrawCommand extends Command {
     public DrawCommand() {
@@ -12,19 +16,17 @@ public class DrawCommand extends Command {
     }
 
     @Override
-    public void execute(String[] commands) {
-        if (commands.length == 1) {
-            Command.sendMessage("Напиши название модуля");
-            return;
-        }
-        String moduleName = commands[0];
-        Module module = Thunderhack.moduleManager.get(moduleName);
-        if (module == null) {
-            Command.sendMessage("Неизвестный модуль'" + module + "'!");
-            return;
-        }
-        if(module.getDisplayName().toLowerCase().contains("click")) mc.currentScreen = ClickUI.getClickGui();
-        module.setDrawn(!module.isDrawn());
-        BindCommand.sendMessage("Модуль " + Formatting.GREEN + module.getName() + Formatting.WHITE + " теперь " + (module.isDrawn() ? "виден в ArrayList" : "не виден в ArrayList"));
+    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
+        builder.then(arg("module", ModuleArgumentType.create()).executes(context -> {
+            Module module = context.getArgument("module", Module.class);
+
+            if (module.getDisplayName().toLowerCase().contains("click"))
+                MC.currentScreen = ClickUI.getClickGui();
+
+            module.setDrawn(!module.isDrawn());
+            sendMessage("Модуль " + Formatting.GREEN + module.getName() + Formatting.WHITE + " теперь " + (module.isDrawn() ? "виден в ArrayList" : "не виден в ArrayList"));
+
+            return SINGLE_SUCCESS;
+        }));
     }
 }
