+++ src/main/java/thunder/hack/cmd/impl/FriendCommand.java
@@ -1,7 +1,13 @@
 package thunder.hack.cmd.impl;
 
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.minecraft.command.CommandSource;
 import thunder.hack.Thunderhack;
 import thunder.hack.cmd.Command;
+import thunder.hack.cmd.args.FriendArgumentType;
+
+import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 
 public class FriendCommand extends Command {
     public FriendCommand() {
@@ -9,46 +15,51 @@ public class FriendCommand extends Command {
     }
 
     @Override
-    public void execute(String[] commands) {
-        if (commands.length == 1) {
+    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
+        builder.then(literal("reset").executes(context -> {
+            Thunderhack.friendManager.clear();
+            sendMessage("Friends got reset.");
+
+            return SINGLE_SUCCESS;
+        }));
+
+        builder.then(literal("add").then(arg("player", StringArgumentType.word()).executes(context -> {
+            String nickname = context.getArgument("player", String.class);
+
+            Thunderhack.friendManager.addFriend(nickname);
+            sendMessage(nickname + " has been friended");
+            return SINGLE_SUCCESS;
+        })));
+
+        builder.then(literal("remove").then(arg("player", FriendArgumentType.create()).executes(context -> {
+            String nickname = context.getArgument("player", String.class);
+
+            Thunderhack.friendManager.removeFriend(nickname);
+            sendMessage(nickname + " has been unfriended");
+            return SINGLE_SUCCESS;
+        })));
+
+        builder.then(arg("player", StringArgumentType.word()).executes(context -> {
+            String nickname = context.getArgument("player", String.class);
+            sendMessage(nickname + (Thunderhack.friendManager.isFriend(nickname) ? " is friended." : " isn't friended."));
+
+            return SINGLE_SUCCESS;
+        }));
+
+        builder.executes(context -> {
             if (Thunderhack.friendManager.getFriends().isEmpty()) {
-                FriendCommand.sendMessage("Friend list empty D:.");
+                thunder.hack.cmd.impl.FriendCommand.sendMessage("Friend list empty D:.");
             } else {
-                String f = "Friends: ";
+                StringBuilder f = new StringBuilder("Friends: ");
                 for (String friend : Thunderhack.friendManager.getFriends()) {
                     try {
-                        f = f + friend + ", ";
+                        f.append(friend).append(", ");
                     } catch (Exception ignored) {
                     }
                 }
-                FriendCommand.sendMessage(f);
+                thunder.hack.cmd.impl.FriendCommand.sendMessage(f.toString());
             }
-            return;
-        }
-        if (commands.length == 2) {
-            if (commands[0].equals("reset")) {
-                Thunderhack.friendManager.clear();
-                FriendCommand.sendMessage("Friends got reset.");
-                return;
-            }
-            FriendCommand.sendMessage(commands[0] + (Thunderhack.friendManager.isFriend(commands[0]) ? " is friended." : " isn't friended."));
-            return;
-        }
-        if (commands.length >= 2) {
-            switch (commands[0]) {
-                case "add" -> {
-                    Thunderhack.friendManager.addFriend(commands[1]);
-                    FriendCommand.sendMessage(commands[1] + " has been friended");
-                    return;
-                }
-                case "del" -> {
-                    Thunderhack.friendManager.removeFriend(commands[1]);
-                    FriendCommand.sendMessage(commands[1] + " has been unfriended");
-                    return;
-                }
-            }
-            FriendCommand.sendMessage("Unknown Command, try friend add/del (name)");
-        }
+            return SINGLE_SUCCESS;
+        });
     }
 }
-
