+++ src/main/java/thunder/hack/cmd/impl/MacroCommand.java
@@ -1,30 +1,29 @@
 package thunder.hack.cmd.impl;
 
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.minecraft.client.util.InputUtil;
+import net.minecraft.command.CommandSource;
+import org.lwjgl.glfw.GLFW;
 import thunder.hack.Thunderhack;
 import thunder.hack.cmd.Command;
+import thunder.hack.cmd.args.MacroArgumentType;
 import thunder.hack.core.MacroManager;
 import thunder.hack.modules.client.MainSettings;
 import thunder.hack.utility.Macro;
-import net.minecraft.client.util.InputUtil;
-import org.lwjgl.glfw.GLFW;
 
 import java.lang.reflect.Field;
-import java.util.Arrays;
+
+import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 
 public class MacroCommand extends Command {
-
-
     public MacroCommand() {
-        super("macro");
+        super("macro", "macros");
     }
 
     @Override
-    public void execute(String[] args) {
-        if (args[0] == null) {
-            Command.sendMessage(usage());
-        }
-        if (args[0].equals("list")) {
-
+    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
+        builder.then(literal("list").executes(context -> {
             if (MainSettings.language.getValue() == MainSettings.Language.RU) {
                 sendMessage("Макросы:");
             } else {
@@ -34,36 +33,53 @@ public class MacroCommand extends Command {
 
             sendMessage(" ");
             Thunderhack.macroManager.getMacros().forEach(macro -> sendMessage(macro.getName() + (macro.getBind() != -1 ? " [" + toString(macro.getBind()) + "]" : "") + " {" + macro.getText() + "}"));
-        }
-        if (args[0].equals("remove")) {
-            if (Thunderhack.macroManager.getMacroByName(args[1]) != null) {
-                Macro macro = Thunderhack.macroManager.getMacroByName(args[1]);
-                Thunderhack.macroManager.removeMacro(macro);
-                sendMessage("Удален макрос " + macro.getName());
-            } else {
-                sendMessage("Не существует макроса с именем " + args[1]);
+
+            return SINGLE_SUCCESS;
+        }));
+
+        builder.then(literal("remove").then(arg("macro", MacroArgumentType.create()).executes(context -> {
+            Macro macro = context.getArgument("macro", Macro.class);
+            if (macro == null) {
+                sendMessage("Не существует такого макроса!");
+                return SINGLE_SUCCESS;
             }
-        }
-        if (args.length >= 4) {
-            if (args[0].equals("add")) {
-                String name = args[1];
-                String bind = args[2].toUpperCase();
-                String text = String.join(" ", Arrays.copyOfRange(args, 3, args.length - 1));
-                if (InputUtil.fromTranslationKey("key.keyboard." + bind.toLowerCase()).getCode() == -1) {
-                    sendMessage("Неправильный бинд!");
-                    return;
-                }
-                Macro macro = new Macro(name, text, InputUtil.fromTranslationKey("key.keyboard." + bind.toLowerCase()).getCode());
-                MacroManager.addMacro(macro);
-                sendMessage("Добавлен макрос " + name + " на кнопку " + toString(macro.getBind() + 0));
-            } else {
-                sendMessage(usage());
-            }
-        }
+
+            Thunderhack.macroManager.removeMacro(macro);
+            sendMessage("Удален макрос " + macro.getName());
+
+            return SINGLE_SUCCESS;
+        })));
+
+        builder.then(literal("add")
+                .then(arg("name", StringArgumentType.word())
+                        .then(arg("bind", StringArgumentType.word())
+                                .then(arg("args", StringArgumentType.string()).executes(context -> {
+                                    String name = context.getArgument("name", String.class);
+                                    String bind = context.getArgument("bind", String.class).toUpperCase();
+                                    String args = context.getArgument("args", String.class);
+
+                                    if (InputUtil.fromTranslationKey("key.keyboard." + bind.toLowerCase()).getCode() == -1) {
+                                        sendMessage("Неправильный бинд!");
+                                        return SINGLE_SUCCESS;
+                                    }
+
+                                    Macro macro = new Macro(name, args, InputUtil.fromTranslationKey("key.keyboard." + bind.toLowerCase()).getCode());
+                                    MacroManager.addMacro(macro);
+                                    sendMessage("Добавлен макрос " + name + " на кнопку " + toString(macro.getBind()));
+
+                                    return SINGLE_SUCCESS;
+                                })))));
+
+        builder.executes(context -> {
+            sendMessage(usage());
+
+            return SINGLE_SUCCESS;
+        });
     }
 
     public String toString(int key) {
-        String kn = key > 0 ? GLFW.glfwGetKeyName((int) (key + 0), GLFW.glfwGetKeyScancode((int) (key + 0))) : "None";
+        String kn = key > 0 ? GLFW.glfwGetKeyName(key, GLFW.glfwGetKeyScancode(key)) : "None";
+
         if (kn == null) {
             try {
                 for (Field declaredField : GLFW.class.getDeclaredFields()) {
@@ -76,7 +92,7 @@ public class MacroCommand extends Command {
                     }
                 }
             } catch (Exception ignored) {
-                kn = "unknown." + (int) (key + 0);
+                kn = "unknown." + key;
             }
         }
