+++ src/main/java/thunder/hack/cmd/impl/WayPointCommand.java
@@ -1,84 +1,84 @@
 package thunder.hack.cmd.impl;
 
+import com.mojang.brigadier.arguments.IntegerArgumentType;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.minecraft.command.CommandSource;
+import net.minecraft.util.math.BlockPos;
 import thunder.hack.Thunderhack;
 import thunder.hack.cmd.Command;
+import thunder.hack.cmd.args.WayPointArgumentType;
 import thunder.hack.core.WayPointManager;
 import thunder.hack.modules.client.MainSettings;
 
+import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
+
 public class WayPointCommand extends Command {
-
-
     public WayPointCommand() {
-        super("waypoint");
+        super("waypoint", "waypoints");
     }
 
     @Override
-    public void execute(String[] args) {
-        if (args[0] == null) {
-            Command.sendMessage(usage());
-        }
-        if (args[0].equals("list")) {
-            if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                sendMessage("Метки:");
-            } else {
-                sendMessage("WayPoints:");
-            }
+    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
+        builder.then(literal("list").executes(context -> {
+            if (MainSettings.language.getValue() == MainSettings.Language.RU) sendMessage("Метки:");
+            else sendMessage("WayPoints:");
+
             sendMessage(" ");
             Thunderhack.wayPointManager.getWayPoints().forEach(wp -> sendMessage(wp.name() + "X: " + wp.x() + " Y: " + wp.y() + " Z: " + wp.z()));
-        }
-        if (args[0].equals("remove")) {
-            if (Thunderhack.wayPointManager.getWayPointByName(args[1]) != null) {
-                WayPointManager.WayPoint wp = Thunderhack.wayPointManager.getWayPointByName(args[1]);
-                Thunderhack.wayPointManager.removeWayPoint(wp);
 
-                if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                    sendMessage("Удалена метка " + wp.name());
-                } else {
-                    sendMessage("Removed waypoint " + wp.name());
-                }
+            return SINGLE_SUCCESS;
+        }));
 
+        builder.then(literal("remove").then(arg("name", WayPointArgumentType.create()).executes(context -> {
+            WayPointManager.WayPoint wp = context.getArgument("name", WayPointManager.WayPoint.class);
+
+            Thunderhack.wayPointManager.removeWayPoint(wp);
+
+            if (MainSettings.language.getValue() == MainSettings.Language.RU) {
+                sendMessage("Удалена метка " + wp.name());
             } else {
-                if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                    sendMessage("Не существует метки с именем " + args[1]);
-                } else {
-                    sendMessage("Waypoint with name " + args[1] + " is not exists");
-                }
+                sendMessage("Removed waypoint " + wp.name());
             }
-        }
-        if (args.length >= 4) {
-            if (args[0].equals("add")) {
-                String x = args[1];
-                String y = args[2];
-                String z = args[3];
-                String name = args[4];
 
-                WayPointManager.WayPoint wp = new WayPointManager.WayPoint (Integer.parseInt(x), Integer.parseInt(y), Integer.parseInt(z), name);
-                Thunderhack.wayPointManager.addWayPoint(wp);
+            return SINGLE_SUCCESS;
+        })));
 
-                if(MainSettings.language.getValue() == MainSettings.Language.RU) {
-                    sendMessage("Добавлена метка " + name + " с координатами x: " + x + " y: " + y + " z: " + z);
-                } else {
-                    sendMessage("Added waypoint " + name + " with coords x: " + x + " y: " + y + " z: " + z);
-                }
+        builder.then(literal("add").then(arg("name", StringArgumentType.word()).executes(context -> {
+            String name = context.getArgument("name", String.class);
+            WayPointManager.WayPoint wp = new WayPointManager.WayPoint((int) MC.player.getX(), (int) MC.player.getY(), (int) MC.player.getZ(), name);
+            Thunderhack.wayPointManager.addWayPoint(wp);
+
+            if (MainSettings.language.getValue() == MainSettings.Language.RU) {
+                sendMessage("Добавлена метка " + name + " с координатами x: " + ((int) MC.player.getX()) + " y: " + ((int) MC.player.getY()) + " z: " + ((int) MC.player.getZ()));
             } else {
-                sendMessage(usage());
+                sendMessage("Added waypoint " + name + " with coords x: " + ((int) MC.player.getX()) + " y: " + ((int) MC.player.getY()) + " z: " + ((int) MC.player.getZ()));
             }
-        } else if(args.length == 3){
-            if (args[0].equals("add")) {
-                String name = args[1];
-                WayPointManager.WayPoint wp = new WayPointManager.WayPoint((int) mc.player.getX(), (int) mc.player.getY(), (int) mc.player.getZ(), name);
-                Thunderhack.wayPointManager.addWayPoint(wp);
-                if (MainSettings.language.getValue() == MainSettings.Language.RU) {
-                    sendMessage("Добавлена метка " + name + " с координатами x: " + ((int) mc.player.getX()) + " y: " + ((int) mc.player.getY()) + " z: " + ((int) mc.player.getZ()));
-                } else {
-                    sendMessage("Added waypoint " + name + " with coords x: " + ((int) mc.player.getX()) + " y: " + ((int) mc.player.getY()) + " z: " + ((int) mc.player.getZ()));
-                }
-            } else {
-                sendMessage(usage());
-            }
-        } else {
+
+            return SINGLE_SUCCESS;
+        }).then(arg("x", IntegerArgumentType.integer())
+                .then(arg("y", IntegerArgumentType.integer())
+                        .then(arg("z", IntegerArgumentType.integer()).executes(context -> {
+                            String name = context.getArgument("name", String.class);
+                            BlockPos pos = new BlockPos(context.getArgument("x", Integer.class), context.getArgument("y", Integer.class), context.getArgument("z", Integer.class));
+
+                            WayPointManager.WayPoint wp = new WayPointManager.WayPoint(pos.getX(), pos.getY(), pos.getZ(), name);
+                            Thunderhack.wayPointManager.addWayPoint(wp);
+
+                            if (MainSettings.language.getValue() == MainSettings.Language.RU) {
+                                sendMessage("Добавлена метка " + name + " с координатами x: " + pos.getX() + " y: " + pos.getY() + " z: " + pos.getZ());
+                            } else {
+                                sendMessage("Added waypoint " + name + " with coords x: " + pos.getX() + " y: " + pos.getY() + " z: " + pos.getZ());
+                            }
+
+                            return SINGLE_SUCCESS;
+                        })))))); // macro remove. // eclip bedrock // waypoints
+
+        builder.executes(context -> {
             sendMessage(usage());
-        }
+
+            return SINGLE_SUCCESS;
+        });
     }
 
     String usage() {
