--- src/main/java/thunder/hack/command/impl/BindCommand.java
@@ -1,61 +1,0 @@
-package thunder.hack.command.impl;
-
-import com.mojang.brigadier.arguments.StringArgumentType;
-import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.minecraft.client.util.InputUtil;
-import net.minecraft.command.CommandSource;
-import net.minecraft.util.Formatting;
-import thunder.hack.command.Command;
-import thunder.hack.command.args.ModuleArgumentType;
-import thunder.hack.modules.Module;
-import thunder.hack.setting.impl.Bind;
-
-import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
-
-public class BindCommand extends Command {
-    public BindCommand() {
-        super("bind");
-    }
-
-    @Override
-    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
-        builder.then(arg("module", ModuleArgumentType.create())
-                .then(arg("key", StringArgumentType.word()).executes(context -> {
-                    final Module module = context.getArgument("module", Module.class);
-                    final String stringKey = context.getArgument("key", String.class);
-
-                    if (stringKey == null) {
-                        sendMessage(module.getName() + " is bound to " + Formatting.GRAY + module.getBind().getBind());
-                        return SINGLE_SUCCESS;
-                    }
-
-                    int key;
-                    if (stringKey.equalsIgnoreCase("none") || stringKey.equalsIgnoreCase("null")) {
-                        key = -1;
-                    } else {
-                        try {
-                            key = InputUtil.fromTranslationKey("key.keyboard." + stringKey.toLowerCase()).getCode();
-                        } catch (NumberFormatException e) {
-                            sendMessage("Такой кнопки не существует!");
-                            return SINGLE_SUCCESS;
-                        }
-                    }
-
-
-                    if (key == 0) {
-                        sendMessage("Unknown key '" + stringKey + "'!");
-                        return SINGLE_SUCCESS;
-                    }
-                    if (!stringKey.equals("M") && stringKey.contains("M")) {
-                        module.bind.setValue(new Bind(key, true));
-                    } else {
-                        module.bind.setValue(new Bind(key, false));
-                    }
-
-                    sendMessage("Bind for " + Formatting.GREEN + module.getName() + Formatting.WHITE + " set to " + Formatting.GRAY + stringKey.toUpperCase());
-
-                    return SINGLE_SUCCESS;
-                }))
-        );
-    }
-}
