--- src/main/java/thunder/hack/command/impl/CfgCommand.java
@@ -1,51 +1,0 @@
-package thunder.hack.command.impl;
-
-import com.mojang.brigadier.arguments.StringArgumentType;
-import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.minecraft.command.CommandSource;
-import thunder.hack.Thunderhack;
-import thunder.hack.command.Command;
-
-import java.io.File;
-import java.util.Objects;
-
-import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
-
-public class CfgCommand extends Command {
-    public CfgCommand() {
-        super("cfg");
-    }
-
-    @Override
-    public void executeBuild(LiteralArgumentBuilder<CommandSource> builder) {
-        builder.then(literal("list").executes(context -> {
-            StringBuilder configs = new StringBuilder("Configs: ");
-            for(String str : Objects.requireNonNull(Thunderhack.configManager.getConfigList())){
-                configs.append("\n- ").append(str);
-            }
-            sendMessage(configs.toString());
-
-            return SINGLE_SUCCESS;
-        }));
-
-        builder.then(literal("dir").executes(context -> {
-            try {
-                net.minecraft.util.Util.getOperatingSystem().open(new File("ThunderHackRecode/configs/").toURI());
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-            return SINGLE_SUCCESS;
-        }));
-
-        builder.then(literal("save").then(arg("name", StringArgumentType.word())).executes(context -> {
-            Thunderhack.configManager.save(context.getArgument("name", String.class));
-            return SINGLE_SUCCESS;
-        }));
-
-        builder.then(literal("load").then(arg("name", StringArgumentType.word()).executes(context -> {
-            Thunderhack.configManager.load(context.getArgument("name", String.class));
-            return SINGLE_SUCCESS;
-        })));
-    }
-}
