+++ src/main/java/thunder/hack/modules/misc/MessageAppend.java
@@ -2,6 +2,7 @@ package thunder.hack.modules.misc;
 
 import com.google.common.eventbus.Subscribe;
 import net.minecraft.network.packet.c2s.play.ChatMessageC2SPacket;
+import thunder.hack.Thunderhack;
 import thunder.hack.cmd.Command;
 import thunder.hack.events.impl.PacketEvent;
 import thunder.hack.injection.accesors.IChatMessageC2SPacket;
@@ -29,7 +30,7 @@ public class MessageAppend extends Module {
             if(Objects.equals(pac.chatMessage(), skip)){
                 return;
             }
-            if (pac.chatMessage().startsWith("/") || pac.chatMessage().startsWith(Command.getCommandPrefix()))
+            if (pac.chatMessage().startsWith("/") || pac.chatMessage().startsWith(Thunderhack.commandManager.getPrefix()))
                 return;
             skip = pac.chatMessage() + word.getValue();
             mc.player.networkHandler.sendChatMessage(pac.chatMessage() + word.getValue());
